About component desktop and mobile
