+++ frontend/debug.log
@@ -18,3 +18,7 @@
 [0729/112618.865:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0729/115823.487:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0730/093728.018:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0730/205431.187:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0731/112103.306:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0731/113846.348:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0731/122240.458:ERROR:crash_report_database_win.cc(469)] failed to stat report
