+++ frontend/src/components/AboutUs/index.js
@@ -1,0 +1,3 @@
+import AboutUs from './AboutUs';
+
+export default AboutUs;
