+++ frontend/src/components/Layout.js
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 
 import Header from './Header';
 import Footer from './Footer';
-import AboutUs from './AboutUs';
 
 // import { above } from '../util';
 
@@ -16,7 +15,6 @@ export default function Layout(props) {
   return (
     <Wrapper>
       <Header />
-      <AboutUs />
       {props.children}
       <Footer />
     </Wrapper>
