+++ frontend/src/elements/ImgBanner.js
@@ -92,4 +92,4 @@ export const ImgMini = styled.img`
   @media (max-width: 600px) {
     display: none;
   }
-  `;
+`;
